+++ docs/releases/releases.md
@@ -1,3 +1,21 @@
+# Release v1.6.3 (27th October 2020)
+
+The release is called **v1.6.3**, an update to support [additional roadmap tests](https://openbanking.atlassian.net/wiki/spaces/DZ/pages/1564083551/OBIE+Functional+Conformance+Tool+Roadmap).
+
+# Release v1.6.2 (9th October 2020)
+
+The release is called **v1.6.2**, an update to support [additional roadmap tests](https://openbanking.atlassian.net/wiki/spaces/DZ/pages/1564083551/OBIE+Functional+Conformance+Tool+Roadmap).
+
+# Release v1.6.1 (23rd September 2020)
+
+The release is called **v1.6.1**, an update to support [additional roadmap tests](https://openbanking.atlassian.net/wiki/spaces/DZ/pages/1564083551/OBIE+Functional+Conformance+Tool+Roadmap). 
+
+# Release v1.6.0 (16th September 2020)
+
+The release is called **v1.6.0**, a minor update that adds support for the Read/Write Data API Specification 3.1.6 including swagger validation for accounts, payments and CBPII. This release is backwards compatible with older versions of the standard.
+
+[Full Release Notes](https://bitbucket.org/openbankingteam/conformance-suite/src/develop/docs/releases/v1.6.0.md) (v1.6.0.md)
+
 # Release v1.5.3 (3rd August 2020)
 
 The release is called **v1.5.3**, a minor update that includes new UI support for SSA & KID for EIDAS certificates, 403 error status code fixes and updates to the report export.